* main: (57 commits)
  chore: release main (#761)
  fix(highlight.run): never JSON.stringify binary XHR bodies or capture the SDK's own exports (iPad crash root cause) (#760)
  chore: release main (#759)
  fix(highlight.run): WebKit/iPad performance hardening — dedupe OpenTelemetry bundles, cap body capture, resync option, telemetry-intake blocklist (#758)
  chore: release main (#754)
  fix(highlight.run): record fetch request bodies and headers that are not plain strings or objects (#753)
  fix(highlight.run): capture XMLHttpRequest bodies and headers on traces without the session replay listener (#752)
  fix: bump next to 15.5.24 / 16.3.3 for Critical GHSA-2xp9-vwfh-vxw4 (#751)
  fix(deps): pin image-size to 1.2.1 in the react-native e2e app (#747)
  chore: release main (#749)
  feat(go): Add explicit ForceSample marker to the span-kind sampler (#738)
  chore: release main (#745)
  fix: bump nanoid, @hono/node-server, and sqlparse to patched versions (#744)
  fix(highlight-go): allow histogram bucket boundaries, and stop racing on the instrument cache (#740)
  chore(react-native): add fading overlay examples to session replay masking tab (#742)
  chore: Update Skip Howdy example (#733)
  chore: release main (#741)
  feat(react-native): fix lagging masks and imageQuality option (#734)
  chore: release main (#736)
  fix(android): anchor session replay masks to the rendered frame and cover stretch overscroll (Rubber animation) (#739)
  ...

+++ sdk/@launchdarkly/observability-android/lib/src/main/kotlin/com/launchdarkly/observability/replay/capture/TiledFrame.kt
@@ -5,7 +5,7 @@ import android.graphics.Bitmap
 data class TiledFrame(
     val id: Int,
     val tiles: List<Tile>,
-    val scale: Float,
+    val scale: Double,
     val originalSize: IntSize,
     val timestamp: Long,
     val orientation: Int,
